% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_type.R
\name{detect_type}
\alias{detect_type}
\title{Detect file type}
\usage{
detect_type(file)
}
\arguments{
\item{file}{A path to a file.}
}
\value{
A lowercase character string giving the detected file extension.
}
\description{
\code{detect_type()} detects a file's type from its extension.
}
\examples{
detect_type("data.csv")
detect_type("survey.DTA")

}
